// src/types/adminType.ts
export interface AdminCourse {
  _id: string;
  course_name: string;
  grade_level: number;
  status: 'active' | 'inactive';
  date_create: string;
  change_log?: any;
}

export interface CreateCourseRequest {
  course_name: string;
  grade_level: number;
}

export interface UpdateCourseRequest {
  course_name?: string;
  grade_level?: number;
  status?: 'active' | 'inactive';
}

export interface AdminTopicCourseInfo {
  course_name: string;
  grade_level: number;
}

export interface AdminTopic {
  _id: string;
  title: string;
  description?: string;
  course_id: string;
  content_json?: any;
  course?: AdminTopicCourseInfo;
}

export interface CreateTopicRequest {
  title: string;
  course_id: string;
  description?: string;
  content_json?: any;
}

export interface UpdateTopicRequest {
  title?: string;
  description?: string;
  content_json?: any;
}